requirements.txt
databases==0.9.0
fastapi==0.115.6
pika==1.3.2
pycurl==7.45.3
pydantic==2.10.4
python-dotenv==1.0.1
SQLAlchemy==2.0.36
SQLAlchemy==1.4.31
uvicorn==0.15.0
